refactor: Rename origin service to runtime-settings
